+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v1.6.0
+
+### Added
+
+- Trait `AppVersionAssertionsTrait`
+
+## v1.5.0
+
+### Added
+
+- `assertArrayStructure` method in `AdditionasAssertionsTrait` [#7]
+- Trait `WithMemoryClean`
+
 ## v1.4.0
 
 ### Added
@@ -91,3 +104,6 @@
 ## v1.0.0
 
 ### First release
+
+
+[#7]:https://github.com/avto-dev/dev-tools/issues/7